Support Gradle Plugin testing with locally-built sigstore-java
The plugin runs with a separate build, so by default it uses mavenCentral only.
However, we need to ensure the plugin is compatible with the main branch,
so we publish sigstore-java to a temp folder, and use it as a repository in tests.

Fixes https://github.com/sigstore/sigstore-java/issues/204

Signed-off-by: Vladimir Sitnikov <sitnikov.vladimir@gmail.com>